requirements.txt
fastapi==0.115.8
kaggle==1.6.17
Pillow==11.1.0
pydantic==2.10.6
pymongo==4.11
python-dotenv==1.0.1
Requests==2.32.3
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
uvicorn
gdown